stream_lit_crack_detetction_git/requirements.txt
absl-py==1.0.0
aiohttp==3.8.0
aiosignal==1.2.0
albumentations==1.1.0
altair==4.2.0
anomalib==0.3.0
antlr4-python3-runtime==4.8
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
boto3==1.20.49
botocore==1.23.49
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
configparser==5.2.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.0.2
Flask-HTTPAuth==4.5.0
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.2.0
fsspec==2022.3.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.0
install==1.3.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.4
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.21.4
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python-headless==4.5.5.62
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathlib2==2.3.7.post1
pathtools==0.1.2
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pyDeprecate==0.3.2
pyee==8.2.2
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytorch-lightning==1.6.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.5.12
shortuuid==1.0.9
six==1.16.0
slackclient==2.9.3
slackeventsapi==3.0.1
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.8.1
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torchmetrics==0.8.2
torchtext==0.12.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
validators==0.18.2
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.13.3
yarl==1.7.2
yaspin==2.1.0
zipfile36==0.1.3
zipp==3.7.0
